refactor(cave-home): isolate reconciliation contracts
